+++ src/linter.ts
@@ -12,9 +12,18 @@ function validateFile(document: vscode.TextDocument) {
     if (config['mainFile'] !== "") {
         fileName = config['mainFile'];
     }
-    child_process.exec(`crystal build --no-color --no-codegen -f json "${fileName}"`, (err, response) => {
-        analyzeDocument(response, '', problemsLimit);
-    })
+    if (document.uri['_formatted'].startsWith('untitled:')) {
+        let child = child_process.spawn('crystal', ['tool', 'format', '--check','-f', 'json', '-']);
+        child.stdin.write(document.getText());
+        child.stdout.on('data', (data) => {
+            analyzeDocument(data.toString(), document.uri, problemsLimit);
+        });
+        child.stdin.end();
+    } else {
+        child_process.exec(`crystal build --no-color --no-codegen -f json "${fileName}"`, (err, response) => {
+            analyzeDocument(response, document.uri, problemsLimit);
+        })
+    }
 }
 
 interface CrystalError {
@@ -25,7 +34,7 @@ interface CrystalError {
     message: string;
 }
 
-export function analyzeDocument(response: string, fileName, problemsLimit = 20) {
+export function analyzeDocument(response: string, uriFormat, problemsLimit = 20) {
     diagnosticCollection.clear();
     let diagnostics = [];
     if (response.startsWith("[{\"file\":\"")) {
@@ -35,7 +44,29 @@ export function analyzeDocument(response: string, fileName, problemsLimit = 20) 
             let problem = results[problemNumber];
             let range = new vscode.Range(problem.line - 1, problem.column - 1, problem.line - 1, (problem.column + (problem.size || 0) - 1));
             let diagnostic = new vscode.Diagnostic(range, problem.message, vscode.DiagnosticSeverity.Error);
-            diagnostics.push([vscode.Uri.file(problem.file || fileName), [diagnostic]]);
+            let file: vscode.Uri;
+            // uriFormat puede ser string o URI
+            if (uriFormat.hasOwnProperty('_formatted')) {
+                if (uriFormat._formatted.startsWith('untitled:')) {
+                    // Formater in untitled
+                    // Live linting (se necesita servidor,
+                    // porque child_process en cada cambio es muy pesado)
+                    // Por ahora no se puede live linter en untitled
+                    file = vscode.Uri.parse(uriFormat);
+                } else {
+                    // Linter para archivo existente en disco
+                    // Se puede live linter with autosaving
+                    // Pueden haber diferentes archivos por lo cual se usa problem.file
+                    // y no document.fileName
+                    file = vscode.Uri.file(problem.file);
+                }
+            } else {
+                // Formater para archivo existente en disco,
+                // se necesita recalcular la uri as string porque crystal
+                // no da información sobre file en tool format
+                file = vscode.Uri.file(uriFormat);
+            }
+            diagnostics.push([file, [diagnostic]]);
         }
     } else {
         return true;